call_center_back/requirements.txt
aiohttp==3.9.5
aiomysql==0.0.22
aiosignal==1.3.1
aiosqlite==0.17.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==2.1.2
frozenlist==1.4.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
openai==0.27.0
PyMySQL==0.9.3
python-dotenv==0.19.2
requests==2.32.3
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==2.1.2
yarl==1.9.4
pandas==1.2.4
